+++ Player.cpp
@@ -1,7 +1,7 @@
 #include "Player.h"
 
 
-Player::Player(float x, float y, SDL_Renderer* renderer)
+Player::Player(int x, int y, SDL_Renderer* renderer)
 {
 	init(x, y, _textureSizeX, _textureSizeY, GameObject::player, _texturePath, renderer, 0, 0);
 }
