+++ Scout.cpp
@@ -1,7 +1,7 @@
 #include "Scout.h"
 
 
-Scout::Scout(float x, float y, SDL_Renderer* screen)
+Scout::Scout(int x, int y, SDL_Renderer* screen)
 {
 	init(x, y, _textureSizeX, _textureSizeY, GameObject::player, _texturePath, screen,_spritestart,0);
 }
